drm/msm/sde: Introduce KCAL color control
Use PCC and HSIC interfaces to manually calibrate the colors of the
DSI panel via sysfs.  This is a simple driver which uses predefined
API to do this. As it is a purely platform driver, sysfs is located
at /sys/devices/platform/kcal_ctrl.0

Supported PCC and HSIC features:
- RGB control,
- Color distortion control,
- Saturation intensity control,
- Contrast intensity control,
- Color brightness control,
- Grayscale mode,
- Astronomy mode;

We already have atomic ops to alter DSI colors, but users prefer
this interface over that.

Signed-off-by: Alesaiko <solcmdr@gmail.com>
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>
Signed-off-by: Nauval Rizky <enuma.alrizky@gmail.com>
Signed-off-by: Mo 森 <rmnscnce@ya.ru>